Update README.md
Here is a learning project; please don't judge the low quality of the work. It is my first day working with anything related to coding. So far, it has been very fun. I am learning fast, thank God. I need to learn a lot. For now, I am an amateur, here for the fun. 
If you like to teach people, if it is your thing, put water in a seed that is just followed on the soil. Please help; I really need some guidance.